Rename crossword package to crossword
Renames the crossword module from `org.churchpresenter.cross` to `org.churchpresenter.crossword` across the app, tests, and UI code. Updates the desktop main class and documentation to match the new package layout so the module launches and references resolve correctly.
