+++ crossword/src/main/kotlin/org/churchpresenter/crossword/ui/AdminApp.kt
@@ -1,4 +1,4 @@
-package org.churchpresenter.cross.ui
+package org.churchpresenter.crossword.ui
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.BoxScope
