+++ crossword/src/main/kotlin/org/churchpresenter/crossword/ui/PuzzleFiles.kt
@@ -1,12 +1,12 @@
-package org.churchpresenter.cross.ui
+package org.churchpresenter.crossword.ui
 
 import java.io.File
-import org.churchpresenter.cross.data.ClueEntry
-import org.churchpresenter.cross.data.CrosswordEngine
-import org.churchpresenter.cross.data.decode
-import org.churchpresenter.cross.data.encode
-import org.churchpresenter.cross.data.fromPlaintext
-import org.churchpresenter.cross.data.toPlaintext
+import org.churchpresenter.crossword.data.ClueEntry
+import org.churchpresenter.crossword.data.CrosswordEngine
+import org.churchpresenter.crossword.data.decode
+import org.churchpresenter.crossword.data.encode
+import org.churchpresenter.crossword.data.fromPlaintext
+import org.churchpresenter.crossword.data.toPlaintext
 
 // Reading the puzzle folder, and the two batch operations that run over everything in it.
